requirements.txt
fastapi
uvicorn
torch
torchvision
timm
numpy
pillow
gdown
